# modern/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Freehub API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/login": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "login",
                  "password"
                ],
                "properties": {
                  "login": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/password": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "minLength": 12,
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/users/{userId}/temporary-password": {
      "post": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 160,
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "maxLength": 320,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/organizations/{organizationId}/settings": {
      "get": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "timezone"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "timezone": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  },
                  "location": {
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/organizations/{organizationId}/memberships": {
      "get": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "role"
                ],
                "properties": {
                  "userId": {
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  },
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "manager"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "operator"
                        ]
                      }
                    ]
                  },
                  "startsAt": {
                    "format": "date",
                    "type": "string"
                  },
                  "endsAt": {
                    "anyOf": [
                      {
                        "format": "date",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/organizations/{organizationId}/memberships/{userId}": {
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role",
                  "startsAt"
                ],
                "properties": {
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "manager"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "operator"
                        ]
                      }
                    ]
                  },
                  "startsAt": {
                    "format": "date",
                    "type": "string"
                  },
                  "endsAt": {
                    "anyOf": [
                      {
                        "format": "date",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          },
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          },
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/organizations": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "key",
                  "timezone",
                  "managerUserId"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 80,
                    "type": "string"
                  },
                  "key": {
                    "minLength": 2,
                    "maxLength": 40,
                    "pattern": "^[a-z0-9-]+$",
                    "type": "string"
                  },
                  "timezone": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  },
                  "location": {
                    "maxLength": 255,
                    "type": "string"
                  },
                  "managerUserId": {
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/organizations/{organizationId}/reports/people.csv": {
      "get": {
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}